decode

+++ rust/src/consts.rs
@@ -2,7 +2,7 @@
 pub const SEPARATOR: char = '+';
 
 // The number of characters to place before the separator.
-pub const SEPARATOR_POSITION: i64 = 8;
+pub const SEPARATOR_POSITION: usize = 8;
 
 // The character used to pad codes.
 pub const PADDING_CHAR: char = '0';
@@ -15,7 +15,7 @@ pub const CODE_ALPHABET: [char; 20] = [
 ];
 
 // The base to use to convert numbers to/from.
-pub const ENCODING_BASE: i64 = 20;
+pub const ENCODING_BASE: usize = 20;
 
 // The maximum value for latitude in degrees.
 pub const LATITUDE_MAX: f64 = 90f64;
@@ -24,15 +24,15 @@ pub const LATITUDE_MAX: f64 = 90f64;
 pub const LONGITUDE_MAX: f64 = 180f64;
 
 // Maximum number of digits to process for plus codes.
-pub const MAX_CODE_LENGTH: i64 = 15;
+pub const MAX_CODE_LENGTH: usize = 15;
 
 // Maximum code length using lat/lng pair encoding. The area of such a
 // code is approximately 13x13 meters (at the equator), and should be suitable
 // for identifying buildings. This excludes prefix and separator characters.
-pub const PAIR_CODE_LENGTH: i64 = 10;
+pub const PAIR_CODE_LENGTH: usize = 10;
 
 // Digits in the grid encoding..
-pub const GRID_CODE_LENGTH: i64 = 5;
+pub const GRID_CODE_LENGTH: usize = 5;
 
 // The resolution values in degrees for each position in the lat/lng pair
 // encoding. These give the place value of each position, and therefore the
@@ -40,10 +40,10 @@ pub const GRID_CODE_LENGTH: i64 = 5;
 pub const PAIR_RESOLUTIONS: [f64; 5] = [20.0f64, 1.0f64, 0.05f64, 0.0025f64, 0.000125f64];
 
 // Number of columns in the grid refinement method.
-pub const GRID_COLUMNS: i64 = 4;
+pub const GRID_COLUMNS: usize = 4;
 
 // Number of rows in the grid refinement method.
-pub const GRID_ROWS: i64 = 5;
+pub const GRID_ROWS: usize = 5;
 
 // Minimum length of a code that can be shortened.
-pub const MIN_TRIMMABLE_CODE_LEN: i64 = 6;
+pub const MIN_TRIMMABLE_CODE_LEN: usize = 6;
